+++ boards/arm/96b_carbon/Makefile
@@ -1,1 +1,8 @@
-obj- += dummy.o
+# Copyright (c) 2017 Linaro Limited
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+
+ccflags-y +=-I$(srctree)/drivers
+
+obj-y += pinmux.o
