+++ boards/arm/nucleo_f411re/pinmux.c
@@ -9,20 +9,19 @@
 #include <init.h>
 #include <pinmux.h>
 #include <sys_io.h>
-#include "pinmux/pinmux.h"
 
-#include "pinmux_stm32.h"
+#include <pinmux/stm32/pinmux_stm32.h>
 
-/* pin assignments for STM32F469I-DISCO board */
+/* pin assignments for NUCLEO-F411RE board */
 static const struct pin_config pinconf[] = {
-#ifdef CONFIG_UART_STM32_PORT_3
-	{STM32_PIN_PB10, STM32F4_PINMUX_FUNC_PB10_USART3_TX},
-	{STM32_PIN_PB11, STM32F4_PINMUX_FUNC_PB11_USART3_RX},
-#endif	/* CONFIG_UART_STM32_PORT_3 */
-#ifdef CONFIG_UART_STM32_PORT_6
-	{STM32_PIN_PG14, STM32F4_PINMUX_FUNC_PG14_USART6_TX},
-	{STM32_PIN_PG9, STM32F4_PINMUX_FUNC_PG9_USART6_RX},
-#endif	/* CONFIG_UART_STM32_PORT_6 */
+#ifdef CONFIG_UART_STM32_PORT_1
+	{STM32_PIN_PB6, STM32F4_PINMUX_FUNC_PB6_USART1_TX},
+	{STM32_PIN_PB7, STM32F4_PINMUX_FUNC_PB7_USART1_RX},
+#endif	/* CONFIG_UART_STM32_PORT_1 */
+#ifdef CONFIG_UART_STM32_PORT_2
+	{STM32_PIN_PA2, STM32F4_PINMUX_FUNC_PA2_USART2_TX},
+	{STM32_PIN_PA3, STM32F4_PINMUX_FUNC_PA3_USART2_RX},
+#endif	/* CONFIG_UART_STM32_PORT_2 */
 };
 
 static int pinmux_stm32_init(struct device *port)
