+++ boards/arm/nucleo_l476rg/pinmux.c
@@ -10,9 +10,8 @@
 #include <init.h>
 #include <pinmux.h>
 #include <sys_io.h>
-#include "pinmux/pinmux.h"
 
-#include "pinmux_stm32.h"
+#include <pinmux/stm32/pinmux_stm32.h>
 
 /* pin assignments for NUCLEO-L476RG board */
 static const struct pin_config pinconf[] = {
