+++ boards/arm/olimexino_stm32/pinmux.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, embedjournal.com
+ * Copyright (c) 2016 Open-RnD Sp. z o.o.
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -9,16 +9,15 @@
 #include <init.h>
 #include <pinmux.h>
 #include <sys_io.h>
-#include "pinmux/pinmux.h"
 
-#include "pinmux_stm32.h"
+#include <pinmux/stm32/pinmux_stm32.h>
 
-/* pin assignments for STM32_MIN_DEV board */
+/* pin assignments for NUCLEO-F103RB board */
 static const struct pin_config pinconf[] = {
 #ifdef CONFIG_UART_STM32_PORT_1
 	{STM32_PIN_PA9,  STM32F1_PINMUX_FUNC_PA9_USART1_TX},
 	{STM32_PIN_PA10, STM32F1_PINMUX_FUNC_PA10_USART1_RX},
-#endif	/* CONFIG_UART_STM32_PORT_1 */
+#endif	/* CONFIG_UART_STM32_PORT_2 */
 #ifdef CONFIG_UART_STM32_PORT_2
 	{STM32_PIN_PA2, STM32F1_PINMUX_FUNC_PA2_USART2_TX},
 	{STM32_PIN_PA3, STM32F1_PINMUX_FUNC_PA3_USART2_RX},
@@ -27,6 +26,10 @@ static const struct pin_config pinconf[] = {
 	{STM32_PIN_PB10, STM32F1_PINMUX_FUNC_PB10_USART3_TX},
 	{STM32_PIN_PB11, STM32F1_PINMUX_FUNC_PB11_USART3_RX},
 #endif	/* CONFIG_UART_STM32_PORT_3 */
+#ifdef CONFIG_I2C_2
+	{STM32_PIN_PB10, STM32F1_PINMUX_FUNC_PB10_I2C2_SCL},
+	{STM32_PIN_PB11, STM32F1_PINMUX_FUNC_PB11_I2C2_SDA},
+#endif /* CONFIG_I2C_2 */
 };
 
 static int pinmux_stm32_init(struct device *port)
