+++ boards/arm/stm32_min_dev/Makefile
@@ -1,2 +1,8 @@
-# No C files (yet)
-obj- += dummy.o
+# Copyright (c) 2017 Linaro Limited
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+
+ccflags-y +=-I$(srctree)/drivers
+
+obj-y += pinmux.o
