+++ boards/arm/stm32f469i_disco/pinmux.c
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2016 Linaro Limited.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <kernel.h>
+#include <device.h>
+#include <init.h>
+#include <pinmux.h>
+#include <sys_io.h>
+
+#include <pinmux/stm32/pinmux_stm32.h>
+
+/* pin assignments for STM32F469I-DISCO board */
+static const struct pin_config pinconf[] = {
+#ifdef CONFIG_UART_STM32_PORT_3
+	{STM32_PIN_PB10, STM32F4_PINMUX_FUNC_PB10_USART3_TX},
+	{STM32_PIN_PB11, STM32F4_PINMUX_FUNC_PB11_USART3_RX},
+#endif	/* CONFIG_UART_STM32_PORT_3 */
+#ifdef CONFIG_UART_STM32_PORT_6
+	{STM32_PIN_PG14, STM32F4_PINMUX_FUNC_PG14_USART6_TX},
+	{STM32_PIN_PG9, STM32F4_PINMUX_FUNC_PG9_USART6_RX},
+#endif	/* CONFIG_UART_STM32_PORT_6 */
+};
+
+static int pinmux_stm32_init(struct device *port)
+{
+	ARG_UNUSED(port);
+
+	stm32_setup_pins(pinconf, ARRAY_SIZE(pinconf));
+
+	return 0;
+}
+
+SYS_INIT(pinmux_stm32_init, PRE_KERNEL_1,
+		CONFIG_PINMUX_STM32_DEVICE_INITIALIZATION_PRIORITY);
