+++ boards/arm/stm32f4_disco/pinmux.c
@@ -9,11 +9,10 @@
 #include <init.h>
 #include <pinmux.h>
 #include <sys_io.h>
-#include "pinmux/pinmux.h"
 
-#include "pinmux_stm32.h"
+#include <pinmux/stm32/pinmux_stm32.h>
 
-/* pin assignments for NUCLEO-F411RE board */
+/* pin assignments for STM32F4DISCOVERY board */
 static const struct pin_config pinconf[] = {
 #ifdef CONFIG_UART_STM32_PORT_1
 	{STM32_PIN_PB6, STM32F4_PINMUX_FUNC_PB6_USART1_TX},
