+++ boards/arm/stm32l496g_disco/pinmux.c
@@ -10,33 +10,26 @@
 #include <init.h>
 #include <pinmux.h>
 #include <sys_io.h>
-#include "pinmux/pinmux.h"
 
-#include "pinmux_stm32.h"
+#include <pinmux/stm32/pinmux_stm32.h>
 
-/* pin assignments for NUCLEO-L432KC board */
+/* pin assignments for STM32L476G-DISCO board */
 static const struct pin_config pinconf[] = {
 #ifdef CONFIG_UART_STM32_PORT_1
-	{STM32_PIN_PA9,  STM32L4X_PINMUX_FUNC_PA9_USART1_TX},
-	{STM32_PIN_PA10, STM32L4X_PINMUX_FUNC_PA10_USART1_RX},
+	{STM32_PIN_PB6,  STM32L4X_PINMUX_FUNC_PB6_USART1_TX},
+	{STM32_PIN_PG10, STM32L4X_PINMUX_FUNC_PG10_USART1_RX},
 #endif	/* CONFIG_UART_STM32_PORT_1 */
 #ifdef CONFIG_UART_STM32_PORT_2
 	{STM32_PIN_PA2, STM32L4X_PINMUX_FUNC_PA2_USART2_TX},
-	{STM32_PIN_PA15, STM32L4X_PINMUX_FUNC_PA15_USART2_RX},
+	{STM32_PIN_PD6, STM32L4X_PINMUX_FUNC_PD6_USART2_RX},
 #endif	/* CONFIG_UART_STM32_PORT_2 */
 #ifdef CONFIG_I2C_1
-	{STM32_PIN_PB6, STM32L4X_PINMUX_FUNC_PB6_I2C1_SCL},
+	{STM32_PIN_PB8, STM32L4X_PINMUX_FUNC_PB8_I2C1_SCL},
 	{STM32_PIN_PB7, STM32L4X_PINMUX_FUNC_PB7_I2C1_SDA},
 #endif /* CONFIG_I2C_1 */
 #ifdef CONFIG_PWM_STM32_2
 	{STM32_PIN_PA0, STM32L4X_PINMUX_FUNC_PA0_PWM2_CH1},
 #endif /* CONFIG_PWM_STM32_2 */
-#ifdef CONFIG_SPI_1
-	{STM32_PIN_PA4, STM32L4X_PINMUX_FUNC_PA4_SPI1_NSS},
-	{STM32_PIN_PA5, STM32L4X_PINMUX_FUNC_PA5_SPI1_SCK},
-	{STM32_PIN_PA6, STM32L4X_PINMUX_FUNC_PA6_SPI1_MISO},
-	{STM32_PIN_PA7, STM32L4X_PINMUX_FUNC_PA7_SPI1_MOSI},
-#endif /* CONFIG_SPI_1 */
 };
 
 static int pinmux_stm32_init(struct device *port)
